% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathtord.R
\name{math_to_rd}
\alias{math_to_rd}
\title{Insert math in R documentation}
\usage{
math_to_rd(tex, ascii = tex, include_css = TRUE, displayMode = TRUE, ...)
}
\arguments{
\item{tex}{string with latex math expression}

\item{ascii}{alternate text representation of the input math to show when documentation is
rendered to plain text format.}

\item{include_css}{adds the katex css file to the output.
This is only required once per html webpage.}

\item{displayMode}{render math in a large, 2D, centered style similar to \verb{$$} in tex.
Set to \code{FALSE} to render inline, which disables centering and tries to squeeze the
equation on a single line. For pdf output, this corresponds to \verb{\deqn{}} and \verb{\eqn{}}
respectively, see \href{https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Mathematics}{WRE 2.6: Mathematics}}

\item{...}{additional rendering options passed to
\href{https://katex.org/docs/options.html}{katex.render}}
}
\description{
Helper function to insert tex math expressions into R documentation (rd) files.
Uses Katex rendering for documentation rendered into HTML format, and the appropriate
tex macros for documentation rendered to pdf
}
\details{
Use \code{math_to_rd()} inside \verb{\\Sexpr} to embed math in your R package documentation
pages. For example the code below:\preformatted{\\Sexpr[results=rd, stage=build]\{
  katex::math_to_rd(katex::example_math())
\}
}

Results in the following output:

\Sexpr[results=rd, stage=build]{
katex::math_to_rd(katex::example_math())
}

Optionally you can specify an ascii representation that will be shown when rendering
the documentation to plain text format:\preformatted{\\Sexpr[results=rd, stage=build]\{
  katex::math_to_rd('E=MC^2', 'E=mc²')
\}
}

\Sexpr[results=rd, stage=build]{
katex::math_to_rd('E=MC^2', 'E=MC²')
}

If no ascii representation is given, the input latex in inserted verbatim into the
plain-text documentation.
}
\seealso{
Other katex: 
\code{\link{katex}}
}
\concept{katex}
