% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katex.R
\name{render_math_html}
\alias{render_math_html}
\alias{render_math_rd}
\title{Render LaTex Math to HTML or MathML}
\usage{
render_math_html(
  tex,
  preview = FALSE,
  include_css = FALSE,
  options = list(output = "html")
)

render_math_rd(tex, include_css = TRUE, options = NULL)
}
\arguments{
\item{tex}{string with latex math expression}

\item{preview}{open an HTML preview page showing the snipped in the browser}

\item{include_css}{adds the katex css file to the output.
This is only required once per page.}

\item{options}{additional arguments for \href{https://katex.org/docs/options.html}{katex.render}}
}
\description{
Converts LaTeX Math to HTML to embed in manual pages or markdown documents.
}
\details{
The conversion is done in R, hence the resulting HTML can be inserted into an
HTML document without the need for a JavaScript library. Only the katex CSS
file is needed to display the final document.

Below is an example of \code{render_math_rd()} used in this manual page:

\Sexpr[results=rd, stage=build]{
katex::render_math_rd(
r"(f(x)= {\frac{1}{\sigma\sqrt{2\pi}}}e^{- {\frac {1}{2}} (\frac {x-\mu}{\sigma})^2})"
)
}
}
\examples{
math <- r"(f(x)= {\frac{1}{\sigma\sqrt{2\pi}}}e^{- {\frac {1}{2}} (\frac {x-\mu}{\sigma})^2})"
render_math_html(math, preview = interactive())
}
