% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katex.R
\name{katex}
\alias{katex}
\alias{katex_html}
\alias{katex_mathml}
\alias{example_math}
\title{Server-side tex math rendering}
\usage{
katex_html(
  tex,
  include_css = FALSE,
  displayMode = TRUE,
  ...,
  preview = interactive()
)

katex_mathml(tex, displayMode = TRUE, ...)

example_math()
}
\arguments{
\item{tex}{input string with tex math expression.}

\item{displayMode}{render math in a large, 2D, centered style similar to \verb{$$} in tex.
Set to \code{FALSE} to render inline, which disables centering and tries to squeeze the
equation on a single line. For pdf output, this corresponds to \verb{\deqn{}} and \verb{\eqn{}}
respectively, see \href{https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Mathematics}{WRE 2.6: Mathematics}}

\item{...}{additional rendering options passed to
\href{https://katex.org/docs/options.html}{katex.render}}

\item{preview}{open an HTML preview page showing the snipped in the browser}
}
\description{
Converts tex-style math expressions to html and mathml for use in manual pages or
markdown documents.
The conversion is to HTML done in R using V8, hence the resulting snipped can
be inserted into an HTML document without the need for a JavaScript library. Only the
\href{https://cdn.jsdelivr.net/npm/katex@0.13.11/dist/katex.min.css}{katex.css}
style is required to display the html in the output document.
Use \link{math_to_rd} for embedding math into R documentation (rd) pages.
}
\details{
By default, \link{katex_html} returns a mix of HTML for visual rendering and includes
MathML for accessibility. To only get html, pass \code{output="html"} in the extra options,
see also the \href{https://katex.org/docs/options.html}{katex documentation}.
}
\examples{
# Basic examples
html <- katex_html(example_math())
mathml <- katex_mathml(example_math())

# Example from katex.org homepage:
macros <- list("\\\\f" = "#1f(#2)")
math <- "\\\\f\\\\relax{x} = \\\\int_{-\\\\infty}^\\\\infty \\\\f\\\\hat\\\\xi\\\\,e^{2 \\\\pi i \\\\xi x} \\\\,d\\\\xi"
html <- katex_html(math,  macros = macros)
mathml <- katex_mathml(math,  macros = macros)
}
\seealso{
Other katex: 
\code{\link{math_to_rd}()}
}
\concept{katex}
