% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katex.R
\name{katex}
\alias{katex}
\alias{render_html}
\alias{render_mathml}
\alias{render_rd}
\alias{example_math}
\title{Render LaTex Math to HTML or MathML}
\usage{
render_html(tex, preview = FALSE, include_css = FALSE, options = NULL)

render_mathml(tex, options = NULL)

render_rd(tex, include_css = TRUE, options = NULL)

example_math()
}
\arguments{
\item{tex}{string with latex math expression}

\item{preview}{open an HTML preview page showing the snipped in the browser}

\item{include_css}{adds the katex css file to the output.
This is only required once per html webpage.}

\item{options}{a list with additional rendering options \code{katex.render()}, see:
\href{https://katex.org/docs/options.html}{katex.render}}
}
\description{
Converts LaTeX Math to HTML to embed in manual pages or markdown documents.
}
\details{
The conversion is done in R, hence the resulting HTML can be inserted into an
HTML document without the need for a JavaScript library. Only the
\href{https://cdn.jsdelivr.net/npm/katex@0.13.11/dist/katex.min.css}{katex css} is
required to display the the html in the output document.

Below is an example of \code{render_math_rd()} used in this manual page:

\Sexpr[results=rd, stage=build]{
katex::render_rd(katex::example_math())
}

By default, \link{render_html} returns a mix of HTML for visual rendering and includes
MathML for accessibility. To only get one or the other, set the \code{output}
parameter in the \code{options} list. For this and other options, see the
\href{https://katex.org/docs/options.html}{katex documentation}.
}
\examples{
html <- render_html(example_math(), preview = interactive())
mathml <- render_mathml(example_math())
}
