% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katex.R
\name{katex}
\alias{katex}
\alias{katex_html}
\alias{katex_rd}
\alias{katex_mathml}
\alias{example_math}
\title{Server-side LaTeX Math Rendering}
\usage{
katex_html(
  tex,
  include_css = FALSE,
  displayMode = TRUE,
  ...,
  preview = interactive()
)

katex_rd(tex, include_css = TRUE, displayMode = TRUE, ...)

katex_mathml(tex, displayMode = TRUE, ...)

example_math()
}
\arguments{
\item{tex}{string with latex math expression}

\item{include_css}{adds the katex css file to the output.
This is only required once per html webpage.}

\item{displayMode}{render math in a large, 2D, centered style similar to \verb{$$} in Latex.
Set to \code{FALSE} to render inline style, which disables centering and tries to
squeeze the equation on a single line.}

\item{...}{additional rendering options passed to
\href{https://katex.org/docs/options.html}{katex.render}}

\item{preview}{open an HTML preview page showing the snipped in the browser}
}
\description{
Converts latex math to html, mathml, or rd for use in manual pages or
markdown documents.
The conversion is done in R using V8, hence the resulting HTML can be inserted into an
HTML document without the need for a JavaScript library. Only the
\href{https://cdn.jsdelivr.net/npm/katex@0.13.11/dist/katex.min.css}{katex.css}
style is required to display the html in the output document.
}
\details{
You can use \code{katex_rd()} inside \verb{\\Sexpr} to embed html math in a package manual
pages. For example the code below:\preformatted{\\Sexpr[results=rd, stage=build]\{
  katex::katex_rd(katex::example_math())
\}
}

Results in the following output (visible in HTML documentation only):

\Sexpr[results=rd, stage=build]{
katex::katex_rd(katex::example_math())
}

By default, \link{katex_html} returns a mix of HTML for visual rendering and includes
MathML for accessibility. To only get html, pass \code{output="html"} in the extra options,
see also the \href{https://katex.org/docs/options.html}{katex documentation}.
}
\examples{
# Basic examples
html <- katex_html(example_math())
mathml <- katex_mathml(example_math())

# Example from katex.org homepage:
macros <- list("\\\\f" = "#1f(#2)")
math <- "\\\\f\\\\relax{x} = \\\\int_{-\\\\infty}^\\\\infty \\\\f\\\\hat\\\\xi\\\\,e^{2 \\\\pi i \\\\xi x} \\\\,d\\\\xi"
html <- katex_html(math,  macros = macros)
mathml <- katex_mathml(math,  macros = macros)
}
